forex_historic/requirements.txt
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
python-dotenv==1.0.1
lxml==5.2.1
selenium==4.19.0
webdriver-manager==4.0.1